Creating the new files and sorting the members of the MemoryManager class into other files
Details:
    - New files:
        - configuration.py
        - ghsConfiguration.py
        - mapfileProcessor.py
        - ghsMapfileProcessor.py
        - mapfileRegexes.py
        - ghsMapfileRegexes.py (this was the old mapfileRegexes.py)
